[MAINT] Improved Makefile commands
I added `go vet` and `golint` to the pipeline.
Then, the `run` command was removed as there
was no need for it. I haven't been using it
all that much. I added the default goal at the
top of the file. Finally, I corrected some general
mistakes in the Makefile syntax and variables.